some minor fixes

+++ STD.h
@@ -11,6 +11,6 @@ typedef struct STD_data
     Compute_float* R;       ///< R STD parameter
 } STD_data;
 STD_data* STD_init(const STD_parameters s);
-Compute_float STD_str (const STD_parameters s, const int x, const int y,const unsigned int time,const unsigned int lag, STD_data* const d);
+Compute_float STD_str (const STD_parameters s, const int x, const int y,const unsigned int time,const int16_t lag, STD_data* const d);
 ///calculation of STD strength.  In .h file for inlining (might not be required)
 #endif
